chore(clean): Added proper enum folder
