+++ GameLogic/Models/Enums/RocketBomb.cs
@@ -1,0 +1,17 @@
+﻿namespace GameLogic.Models.Enums
+{
+    public enum RocketBomb
+    {
+        rocket1 = 4,
+        rocket2 = 23,
+        rocket3 = 29,
+        rocket4 = 44,
+        rocket5 = 63,
+        rocket6 = 71,
+        bomb1 = 15,
+        bomb2 = 72,
+        bomb3 = 81,
+        bomb4 = 94,
+        bomb5 = 98
+    }
+}
